Moved fixed node creation into its own module
